+++ include/test/test_ascon_suite.hpp
@@ -5,6 +5,6 @@
 #include "test_ascon80pq_aead.hpp"
 #include "test_ascon_hash.hpp"
 #include "test_ascon_hasha.hpp"
+#include "test_ascon_perm.hpp"
 #include "test_ascon_xof.hpp"
 #include "test_ascon_xofa.hpp"
-#include "test_permutation.hpp"
